Simplify Eve setup and call Shopify directly
