+++ apps/docs/content/docs/anatomy/cart.mdx
@@ -22,7 +22,7 @@ The cart displays confirmed discounts and sale pricing. Shopify bundles remain g
 
 When customer authentication is enabled, sign-in can attach the existing browser cart to the customer. Checkout can then use the customer's saved details. See [Authentication](/docs/anatomy/authentication).
 
-The optional shopping assistant uses the same cart. The storefront refreshes it after the assistant's response settles. Stopping a response does not roll back a Shopify write that has already started. Checkout remains unavailable during the response and cart reconciliation; an unconfirmed write can temporarily block checkout for up to two minutes. Inspect the cart before requesting another change. Restoring an old conversation does not replay its cart changes.
+The shopping assistant uses the same cart. The storefront refreshes it after the assistant's response settles. Stopping a response does not roll back a Shopify write that has already started. Checkout in that tab waits during the response and cart reconciliation. Inspect the cart before requesting another change. The browser does not resubmit mutations from restored messages, but Eve can rerun an interrupted tool step; see [Agent](/docs/anatomy/agent) for replay and access limitations.
 
 If `NEXT_PUBLIC_SHOPIFY_STOREFRONT_ID` is set, server-side cart activity is attributed to the corresponding Headless storefront. See [Environment Variables](/docs/reference/env-vars).
