+++ apps/docs/content/docs/reference/routes.mdx
@@ -26,14 +26,13 @@ type: reference
 | ---------------------------------------------- | --------------------------------------------------------------------------------------------------- |
 | `POST /api/webhooks/shopify`                   | Shopify webhook handler for cache invalidation                                                      |
 | `/eve/v1/session` and `/eve/v1/session/[id]/*` | Eve conversation creation, messages, streams, and controls; available when the assistant is enabled |
-| `POST /api/agent/session`                      | Same-origin assistant browser-session and cart setup                                                |
-| `POST /api/agent/commerce`                     | Signed server-to-server storefront adapter; not a public chat endpoint                              |
+| `POST /api/agent/session`                      | Same-origin cart setup before an assistant message                                                  |
 | `GET /agent/ucp-profile.json`                  | Public assistant profile when enabled; distinct from the merchant's `/.well-known/ucp`              |
 | `GET /api/cart`                                | Current cart                                                                                        |
 | `POST /api/cart`                               | Cart creation and updates                                                                           |
 | `GET /api/predictive-search`                   | Search suggestions                                                                                  |
 
-Cart and predictive-search endpoints use the framework-agnostic Hydrogen preview SDK's registered handlers through the Next.js proxy. They are not App Router `route.ts` files. Next.js owns the page routes, rendering, webhook, and assistant setup/commerce adapters. Eve owns conversation routes on the same origin.
+Cart and predictive-search endpoints use the framework-agnostic Hydrogen preview SDK's registered handlers through the Next.js proxy. They are not App Router `route.ts` files. Next.js owns the page routes, rendering, webhook, and assistant cart setup. Eve owns conversation routes on the same origin and calls Shopify directly.
 
 ## Customer accounts
