+++ apps/template/.env.example
@@ -13,10 +13,5 @@ NEXT_PUBLIC_SHOPIFY_STOREFRONT_ACCESS_TOKEN="your-storefront-access-token-here"
 # SHOPIFY_WEBHOOK_SECRET="your-shopify-webhook-secret-here"
 # UCP_AGENT_PROFILE_URL="https://your-agent.example.com/agent/ucp-profile.json" # Optional override for the public Eve agent profile.
 
-# Required when agent.isEnabled is true. Keep the secret and Redis credentials server-only.
-# AGENT_SESSION_SECRET="your-agent-session-secret-at-least-32-characters-here" # Generate with openssl rand -base64 32.
-# AGENT_REDIS_URL="https://your-redis-instance.upstash.io" # Upstash Redis REST URL for private session ownership and replay protection.
-# AGENT_REDIS_TOKEN="your-redis-rest-token-here" # Upstash Redis REST token for the same database.
-# AGENT_STOREFRONT_URL="http://localhost:3000" # Exact browser-facing origin; set the deployment's HTTPS URL in production and previews.
-# VERCEL_AUTOMATION_BYPASS_SECRET="your-vercel-bypass-secret-here" # Only when deployment protection blocks Eve's same-project commerce requests.
 # EVE_DEV="1" # Set by eve dev; selects Shopify's public example profile when local URLs are unreachable.
+# VERCEL_ENV is provided by Vercel; previews use Shopify's public example agent profile.
