--- apps/template/lib/agent/session/types.ts
@@ -1,10 +1,0 @@
-export interface AgentSessionBinding {
-  cartId: string;
-  shopperId: string;
-}
-
-export interface ShopperSession {
-  cartId: string;
-  expiresAt: number;
-  id: string;
-}
