+++ apps/template/lib/collections/action.ts
@@ -4,7 +4,7 @@ import { PRODUCTS_PER_PAGE } from "@/lib/collections";
 import { resolveBrowseParams } from "@/lib/collections/server";
 import type { PageInfo } from "@/lib/pagination/types";
 import type { ProductCard } from "@/lib/product/types";
-import { fetchCollectionProducts } from "@/lib/shopify/operations/products/server";
+import { fetchCollectionProducts } from "@/lib/shopify/catalog/server";
 
 export async function loadMoreCollectionProductsAction(params: {
   collection: string;
