+++ apps/template/lib/collections/server.ts
@@ -3,11 +3,10 @@ import { parseCollectionParams, serializeCollectionParams } from "@shopify/hydro
 import { getActiveFilters, getCollectionSortFromState } from "@/lib/collections";
 import { PRODUCTS_PER_PAGE } from "@/lib/collections";
 import type { Collection } from "@/lib/collections/types";
+import { fetchCollectionProducts, fetchSearchIndexProducts } from "@/lib/shopify/catalog/server";
 import {
   buildProductFiltersFromParams,
-  fetchCollectionProducts,
   fetchSearchFacets,
-  fetchSearchIndexProducts,
 } from "@/lib/shopify/operations/products/server";
 
 import type { CollectionResultsData, CollectionSearchState } from "./types";
