+++ apps/template/lib/collections/types.ts
@@ -1,7 +1,7 @@
 import type { Filter, PriceRange } from "@/lib/filters/types";
 import type { Image } from "@/lib/media/types";
 import type { SEO } from "@/lib/seo/types";
-import type { fetchCollectionProducts } from "@/lib/shopify/operations/products/server";
+import type { CollectionProductsResult } from "@/lib/shopify/operations/products/types";
 import type { ProductFilter } from "@/lib/shopify/transforms/filters/types";
 
 export type ActiveFilters = Record<string, string | string[] | undefined>;
@@ -19,7 +19,7 @@ export interface CollectionSearchState extends BrowseParams {
 export interface CollectionResultsData extends BrowseParams {
   collection: string;
   dataSearch: string;
-  result: Awaited<ReturnType<typeof fetchCollectionProducts>>;
+  result: CollectionProductsResult;
   transformedFilters: { filters: Filter[]; priceRange?: PriceRange };
 }
