+++ apps/template/lib/search/action.ts
@@ -4,7 +4,7 @@ import { PRODUCTS_PER_PAGE } from "@/lib/collections";
 import { resolveBrowseParams } from "@/lib/collections/server";
 import type { PageInfo } from "@/lib/pagination/types";
 import type { ProductCard } from "@/lib/product/types";
-import { fetchSearchIndexProducts } from "@/lib/shopify/operations/products/server";
+import { fetchSearchIndexProducts } from "@/lib/shopify/catalog/server";
 
 export async function loadMoreSearchProductsAction(params: {
   collection?: string;
