+++ apps/template/lib/search/server.ts
@@ -2,10 +2,8 @@ import { createPredictiveSearchServerHandlers, gql } from "@shopify/hydrogen";
 
 import { PRODUCTS_PER_PAGE } from "@/lib/collections";
 import type { CollectionSearchState } from "@/lib/collections/types";
-import {
-  fetchSearchFacets,
-  fetchSearchIndexProducts,
-} from "@/lib/shopify/operations/products/server";
+import { fetchSearchIndexProducts } from "@/lib/shopify/catalog/server";
+import { fetchSearchFacets } from "@/lib/shopify/operations/products/server";
 
 import type { SearchResultsData } from "./types";
