+++ apps/template/package.json
@@ -28,7 +28,6 @@
     "@types/node": "26.4.1",
     "@types/react": "19.2.18",
     "@types/react-dom": "19.2.7",
-    "@upstash/redis": "1.38.4",
     "@vercel/analytics": "2.0.1",
     "@vercel/speed-insights": "2.0.0",
     "ai": "7.0.93",
